refactor(kafka): isolate replay and align metrics
